// autogpt_platform/frontend/src/app/api/__generated__/models/nodeExecutionResult.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { AgentExecutionStatus } from "./agentExecutionStatus";
import type { NodeExecutionResultInputData } from "./nodeExecutionResultInputData";
import type { NodeExecutionResultOutputData } from "./nodeExecutionResultOutputData";
import type { NodeExecutionResultQueueTime } from "./nodeExecutionResultQueueTime";
import type { NodeExecutionResultStartTime } from "./nodeExecutionResultStartTime";
import type { NodeExecutionResultEndTime } from "./nodeExecutionResultEndTime";

export interface NodeExecutionResult {
  user_id: string;
  graph_id: string;
  graph_version: number;
  graph_exec_id: string;
  node_exec_id: string;
  node_id: string;
  block_id: string;
  status: AgentExecutionStatus;
  input_data: NodeExecutionResultInputData;
  output_data: NodeExecutionResultOutputData;
  add_time: string;
  queue_time: NodeExecutionResultQueueTime;
  start_time: NodeExecutionResultStartTime;
  end_time: NodeExecutionResultEndTime;
}